+++ src/views/common/Modal.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import './modal.css';
+
+const Modal = (props) => {
+  // 열기, 닫기, 모달 헤더 텍스트를 부모로부터 받아옴
+  const { open, close, header } = props;
+
+  return (
+    // 모달이 열릴때 openModal 클래스가 생성된다.
+    <div className={open ? 'openModal modal' : 'modal'}>
+      {open ? (
+        <section>
+          <header>
+            {header}
+            <button className="close" onClick={close}>
+              &times;
+            </button>
+          </header>
+          <main>{props.children}
+          <h5>정말로 탈퇴하시겠습니까?</h5>
+          </main>
+          <footer>
+          <button className="yes" onClick={close}>
+              예
+            </button>  
+            <button className="close" onClick={close}>
+              아니오
+            </button>
+          </footer>
+        </section>
+      ) : null}
+    </div>
+  );
+};
+
+export default Modal;
