+++ src/views/customerService/customerService.css
@@ -1,0 +1,5 @@
+.customerServiceNav {
+   list-style-type: none;
+   float: left;
+   margin-left: 20px;
+}
