+++ src/views/myPage/myPage.css
@@ -1,0 +1,7 @@
+.updateButton{
+    
+}
+
+.deleteButton {
+   
+}
